requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bidict==0.22.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
Flask==1.0.2
Flask-Argon2==0.2.0.0
flask-jwt-login==0.0.5
Flask-SocketIO==5.2.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==1.1.0
Jinja2==2.11.3
markup==0.2
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==1.6.4
python-engineio==4.2.1
python-socketio==5.4.0
requests==2.27.1
urllib3==1.26.9
waitress==2.1.2
Werkzeug==1.0.1
zipp==3.8.0